pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.uv.workspace]
members = [
    "frontend-idea/backend",
]

[project]
name = "book-translator"
version = "0.1.0"
description = "Extract text from PDF books chapter by chapter using Gemini 2.5 Flash"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pypandoc>=1.13",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "mistralai>=1.9.2",
    "requests>=2.32.4",
    "aiohttp>=3.12.14",
    "cerebras-cloud-sdk>=1.35.0",
    "transformers>=4.53.3",
    "tqdm>=4.67.1",
    "alive-progress>=3.3.0",
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "accelerate>=1.9.0",
]

[tool.pyright]
typeCheckingMode = "strict"



[project.scripts]
book-translator = "book_translator.book_processor:main"
create-ebook = "book_translator.create_ebook:main"

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "ruff>=0.12.4",
    "types-Pillow>=10.0.0",
    "pymupdf-stubs>=1.0.0",
]
